requirements.txt
fastapi
uvicorn
requests
python-dotenv
click
pydantic
